backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-jazzmin==3.0.0
sqlparse==0.5.1
tzdata==2024.1
djangorestframework==3.15.1